backups/backup_250903/requirements.txt
Flask==3.0.3
pymongo==4.8.0
python-dotenv==1.0.1
pytz==2024.1
APScheduler==3.10.4
Flask-Cors==4.0.1
PyJWT==2.9.0
Pillow==10.0.0